requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyMuPDF
numpy
requests
python-dotenv
streamlit>=1.30.0
PyMuPDF>=1.23.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rapidfuzz>=3.5.2
requests>=2.31.0
python-dotenv>=1.0.1
numpy>=1.24.0